+++ mobile/android/android-components/components/feature/share/src/test/java/mozilla/components/feature/share/RecentAppStorageTest.kt
@@ -102,4 +102,14 @@ class RecentAppStorageTest {
 
         verify(recentAppsDao).insertRecentApp(result)
     }
+
+    @Test
+    fun `delete an app from our database`() {
+        val deleteAppName = "delete"
+        val recentApp = RecentAppEntity(score = 1.0, activityName = deleteAppName)
+
+        recentAppsStorage.deleteRecentApp(recentApp.activityName)
+
+        verify(recentAppsDao).deleteRecentApp(deleteAppName)
+    }
 }